Reintroduces the TIME_WAIT for ltcei
Reintroduces the TIME_WAIT for ltcei

The linger option is necessary for the local tcp client endpoints,
as when closing a socket it will stay in TIME_WAIT for a default time,
in linux systems, of 120 seconds.
In this change we change the TIME_WAIT to 5 seconds.

Setting the TIME_WAIT to 0 seconds forces RST to always be sent
in response to a FIN. So a normal shutdown would look like this:
[FIN]
[ACK, FIN]
[ACK, RST]

Since this is endpoint for internal communication setting the TIME_WAIT
to 5 seconds should be enough to ensure the final ACK to the FIN arrives
to the server endpoint. With the 5 seconds the shutdown will look like
this:
[FIN]
[ACK, FIN]
[ACK]